fix: update incorrect MIME type default
Audio is always stored in the browser as WAVE PCM audio.

When we export the blob, despite providing an mp3 MIME type, the resulting Blob is still actually a WAVE PCM file. The MIME type provided is just to help the browser make sense of the Blob and does not determine the data exported.

If you open a downloaded Crunker audio file, the magic bytes (`52 49 46 46`) point to the exported file being a `.wav` file. Most software fix this silently in the background for you, so it's unlikely that it would even be noticed.